requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
jinja2>=3.0.0
python-multipart>=0.0.6
httpx>=0.25.0
websockets>=11.0.0
openpyxl>=3.1.0
pandas>=2.0.0
chromadb>=0.4.0
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0
soundfile>=0.12.0
sentence-transformers>=2.2.0
librosa>=0.10.0
scipy>=1.10.0
